% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/county.R
\name{county}
\alias{county}
\title{Get county information from the Spitcast API}
\usage{
county(item, county_name = "")
}
\arguments{
\item{item}{The name of the specific county data item as a string.  Must be
one of must be one of 'spots', 'swell', 'tide', 'water-temperature', or
'wind'. Required.}

\item{county_name}{The county name to query, as a string.  Required.}
}
\description{
Get county information from the Spitcast API
}
\examples{
county(item = "spots", county_name = "Orange County")
county(item = "swell", county_name = "Los Angeles")
}

