% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot.R
\name{spot}
\alias{spot}
\title{Get spot information from the Spitcast API}
\usage{
spot(item = "all", spot_id = NULL, ...)
}
\arguments{
\item{item}{The name of the specific spot data item as a string.  Must must
be one of 'all', 'forecast', 'nearby', 'neighbors', 'top', or 'search'.
Required.  Default is 'all'.}

\item{spot_id}{The spot ID to query, as an integer.  Required for the
'forecast', and 'neighbors' items.  Default is \code{NULL}.}

\item{...}{Any additional parameters passed to specific data items.  For
example, the 'neighbors' item takes an optional 'direction' parameter.}
}
\description{
Get spot information from the Spitcast API
}
\examples{
spot(item = "all")
spot(item = "nearby")
spot(item = "search", size_min = 1, shape_min = 1, distance = 10)
spot(item = "neighbors", spot_id = 1, direction = "above")
spot(item = "forecast", spot_id = 1)
}
